feat(make_idempotent): sync `atomic_idempotent` from meta server into `.app-info` file for each replica (#2220)
https://github.com/apache/incubator-pegasus/issues/2197

In https://github.com/apache/incubator-pegasus/pull/2205, we introduced a
new attribute `atomic_idempotent` to meta, which is used to enable/disable
the idempotence of atomic write operations.

This attribute should be broadcast to each replica: whenever one replica is
promoted to the primary, by this attribute it can decide whether to make all
atomic writes idempotent. This attribute will be persisted into `.app-info`
file to ensure it can be loaded after restarted.