ci(github): fix Go workflows which failed due to Ubuntu 20.04 retirement (#2240)
Resolve https://github.com/apache/incubator-pegasus/issues/2241.

Go workflows run on Ubuntu 20.04 because Go client uses thrift-compiler 0.13.0
to generate code and the thrift-compiler version on Ubuntu 20.04 is 0.13.0. However
Ubuntu 20.04 runner has been retired and removed on 2025-04-15. To solve this
problem, we choose to build thrift 0.13.0 in a image based on Ubuntu 22.04.

Also add Collector to the workflow for regular building.