feat(duplication): support resumable download for checkpoint files during full duplication on the server side (#2238)
https://github.com/apache/incubator-pegasus/issues/2242

When a dup follower requests the latest checkpoint information from the dup
master, the server supports returning the size and checksum of each file under
the checkpoint directory. This allows the client to decide which files to fetch:

- if it is the first time downloading the checkpoint, all files need to be downloaded;
- if it is the second time or later, only files that are missing or partially downloaded
locally need to be fetched.

Additionally, the client is allowed to specify the checksum algorithm to use —
currently, only MD5 is supported.