feat(make_idempotent): re-adapt `make_idempotent()` and `put()` to support batched single-update requests for atomic writes (#2249)
https://github.com/apache/incubator-pegasus/issues/2197

Because a `check_and_mutate` request may be converted into multiple single-update
requests, the `make_idempotent()` and `put()` interfaces for `incr`, `check_and_set`
and `check_and_mutate` have been re-adapted and better supported through template
functions.

The number of batched single-operation requests corresponding to the batched requests
of put/remove, or an atomic write request of incr, check_and_set or check_and_mutate
has been changed from being stored in a variable to being recorded in an array.

In addition, the writers responsible for applying single-update requests to the storage
engine are now stored in an array instead of a map to improve performance.