feat(go-client): generate interfaces for the hotkey detection (#2255)
To enable the Go collector to implement hotspot detection, it is necessary to
generate the hotkey detection interfaces in the Go client for invocation. The files
`go-client/generator/radmin.csv` and `idl/replica_admin.thrift` were modified to
generate the interfaces. All other newly added or modified Go source files are
generated and formatted.