ci(github): pin the action `dorny/paths-filter` to the commit SHA pointing to `v3.0.2` (#2275)
Resolve https://github.com/apache/incubator-pegasus/issues/2276.

According to the the reply from INFRA team[1] and the doc of infrastructure
actions[2], the actions in allow list must be pinned to a SHA rather than a tag
as of August 1st.

Only the action `rebuild_thirdparty_if_needed`[3] uses `dorny/paths-filter`.
Events are also added to the workflows depending on it so that they could be
triggered once it is changed.

1. https://issues.apache.org/jira/browse/INFRA-27084
2. https://github.com/apache/infrastructure-actions/?tab=readme-ov-file#adding-a-new-action-to-the-allow-list
3. https://github.com/apache/incubator-pegasus/blob/master/.github/actions/rebuild_thirdparty_if_needed/action.yaml#L22