fix(zookeeper): fix potential dangling path pointer issue when calling ZooKeeper APIs (#2302)
Fix https://github.com/apache/incubator-pegasus/issues/2307.

The heap-use-after-free error found in the ASan test was caused by a dangling
pointer issue: the `path` pointer was freed in the thread for completion callback
while it was still used in the ZooKeeper C client API.

To fix this problem, `std::shared_ptr` is introduced to manage the `path`, ensuring
that the memory is automatically released only when all references to it have been
destroyed.