refactor: extract and encapsulate the aggregation process of single-table statistics for reuse (#2318)
Both shell and HTTP interfaces now support displaying a table's statistics in
multiple sections. Since the logic for this was largely duplicated, a lot of
redundant code existed. This PR extracts the common logic and introduces
an `add_app_info` interface to encapsulate it, allowing a table’s statistics to
be added to a `multi_printer` in multiple sections for later output.

Additionally, a `resolve` interface is introduced for `host_port` itself, replacing
`replication_ddl_client::node_name`.